+++ example/migrations/0004_auto_20171011_0631.py
@@ -3,7 +3,6 @@ from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("example", "0003_polymorphics"),
     ]