+++ example/migrations/0005_auto_20180922_1508.py
@@ -5,7 +5,6 @@ from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("example", "0004_auto_20171011_0631"),
     ]