+++ example/migrations/0010_auto_20210714_0809.py
@@ -4,7 +4,6 @@ from django.db import migrations
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("example", "0009_labresults_author"),
     ]