+++ example/migrations/0012_author_full_name.py
@@ -4,7 +4,6 @@ from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ("example", "0011_rename_type_author_author_type_and_more"),
     ]