+++ rest_framework_json_api/relations.py
@@ -1,5 +1,4 @@
 import json
-from collections import OrderedDict
 
 import inflection
 from django.core.exceptions import ImproperlyConfigured
@@ -104,7 +103,7 @@ class HyperlinkedMixin:
     def get_links(self, obj=None, lookup_field="pk"):
         request = self.context.get("request", None)
         view = self.context.get("view", None)
-        return_data = OrderedDict()
+        return_data = {}
 
         kwargs = {
             lookup_field: getattr(obj, lookup_field)
@@ -257,7 +256,7 @@ class ResourceRelatedField(HyperlinkedMixin, PrimaryKeyRelatedField):
         if resource_type is None or not self._skip_polymorphic_optimization:
             resource_type = get_resource_type_from_instance(value)
 
-        return OrderedDict([("type", resource_type), ("id", str(pk))])
+        return {"type": resource_type, "id": str(pk)}
 
     def get_resource_type_from_included_serializer(self):
         """
@@ -301,12 +300,10 @@ class ResourceRelatedField(HyperlinkedMixin, PrimaryKeyRelatedField):
         if cutoff is not None:
             queryset = queryset[:cutoff]
 
-        return OrderedDict(
-            [
-                (json.dumps(self.to_representation(item)), self.display_value(item))
-                for item in queryset
-            ]
-        )
+        return {
+            json.dumps(self.to_representation(item)): self.display_value(item)
+            for item in queryset
+        }
 
 
 class PolymorphicResourceRelatedField(ResourceRelatedField):