+++ CHANGELOG.md
@@ -17,10 +17,13 @@ any parts of the framework not mentioned in the documentation should generally b
 ### Changed
 
 * Added support to overwrite serializer methods in customized schema class
+* Adjusted some still old formatted strings to f-strings.
+* Replaced `OrderedDict` with `dict` which is also ordered since Python 3.7.
 
 ### Fixed
 
 * Refactored handling of the `sort` query parameter to fix duplicate declaration in the generated schema definition
+* Non-field serializer errors are given a source.pointer value of "/data".
 * Serialization of non-model `Serializer` results, e.g. `dict` without a `pk` attribute
 
 ## [6.0.0] - 2022-09-24